// src/shared/types/DetailsProductType.ts
export interface DetailsProductType {
  id: string;
  category: string;
  namespaceId: string;
  name: string;
  capacityAvailable: string[];
  capacity: string;
  priceRegular: number;
  priceDiscount: number;
  colorsAvailable: string[];
  color: string;
  images: string[];
  description: { title: string; text: string[] }[];
  screen: string;
  resolution: string;
  processor: string;
  camera: string;
  ram: string;
  zoom: string;
  cell: string[];
}